requirements.txt
streamlit==1.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.6.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
langchain==0.0.340
langchain-community==0.0.1
tiktoken==0.5.2
pdfplumber
